feat: Implement ThemeCubit for managing app themes
